+++ packages/e2e-test-app-fabric/jest.config.js
@@ -16,8 +16,6 @@ const defaultTransform = [
   require('@rnw-scripts/babel-node-config'),
 ];
 
-const appWorkingDir = 'windows\\x64\\Release\\RNTesterApp-Fabric';
-
 module.exports = {
   preset: '@rnx-kit/jest-preset',
 
@@ -62,10 +60,15 @@ module.exports = {
   setupFilesAfterEnv: ['react-native-windows/jest/setup', './jest.setup.js'],
 
   testEnvironmentOptions: {
-    app: `${appWorkingDir}\\RNTesterApp-Fabric.exe`,
-    appWorkingDir: appWorkingDir,
+    app: `windows\\${
+      (process.env.E2ETestFabricBuildPlatform === 'x86'
+        ? ''
+        : (process.env.E2ETestFabricBuildPlatform ?? 'x64') + '\\') +
+      (process.env.E2ETestFabricBuildConfiguration ?? 'Release')
+    }\\RNTesterApp-Fabric.exe`,
+    appWorkingDir: 'windows\\RNTesterApp-Fabric',
     enableAutomationChannel: true,
-    /* -- Enable for more detailed logging
+    /* // Enable for more detailed logging
     webdriverOptions: {
       // Level of logging verbosity: trace | debug | info | warn | error
       logLevel: 'info',