+++ vnext/Microsoft.ReactNative/Fabric/ComponentView.h
@@ -11,6 +11,7 @@
 #include <react/renderer/components/view/ViewProps.h>
 #include <react/renderer/core/LayoutMetrics.h>
 
+#include <Fabric/Composition/Theme.h>
 #include <winrt/Microsoft.ReactNative.Composition.Input.h>
 
 namespace Microsoft::ReactNative {
@@ -23,6 +24,12 @@ enum class RNComponentViewUpdateMask : std::uint_fast8_t {
   LayoutMetrics = 1 << 3,
 
   All = Props | EventEmitter | State | LayoutMetrics
+};
+
+enum class ClipState : std::uint_fast8_t {
+  NoClip = 0,
+  PartialClip = 1,
+  FullyClipped = 2,
 };
 
 DEFINE_ENUM_FLAG_OPERATORS(RNComponentViewUpdateMask);
@@ -62,11 +69,16 @@ struct IComponentView {
   virtual RootComponentView *rootComponentView() noexcept = 0;
   virtual void parent(IComponentView *parent) noexcept = 0;
   virtual IComponentView *parent() const noexcept = 0;
+  virtual void theme(const std::shared_ptr<Composition::Theme> &theme) noexcept = 0;
+  virtual std::shared_ptr<Composition::Theme> &theme() const noexcept = 0;
+  virtual void onThemeChanged() noexcept = 0;
   virtual const std::vector<IComponentView *> &children() const noexcept = 0;
   // Run fn on all children of this node until fn returns true
   // returns true if the fn ever returned true
   virtual bool runOnChildren(bool forward, Mso::Functor<bool(IComponentView &)> &fn) noexcept = 0;
   virtual RECT getClientRect() const noexcept = 0;
+  // The offset from this elements parent to its children (accounts for things like scroll position)
+  virtual facebook::react::Point getClientOffset() const noexcept = 0;
   virtual void onFocusLost() noexcept = 0;
   virtual void onFocusGained() noexcept = 0;
   virtual void onPointerEntered(
@@ -101,8 +113,13 @@ struct IComponentView {
       facebook::react::Point &localPt,
       bool ignorePointerEvents = false) const noexcept = 0;
   virtual winrt::IInspectable EnsureUiaProvider() noexcept = 0;
+  virtual std::optional<std::string> getAcccessiblityValue() noexcept = 0;
+  virtual void setAcccessiblityValue(std::string &&value) noexcept = 0;
+  virtual bool getAcccessiblityIsReadOnly() noexcept = 0;
+
   // Notify up the tree to bring the rect into view by scrolling as needed
   virtual void StartBringIntoView(BringIntoViewOptions &&args) noexcept = 0;
+  virtual ClipState getClipState() noexcept = 0;
 };
 
 // Run fn on all nodes of the component view tree starting from this one until fn returns true