+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionViewComponentView.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <Fabric/ComponentView.h>
+#include <Microsoft.ReactNative.Cxx/ReactContext.h>
 #include <react/renderer/components/view/ViewEventEmitter.h>
 #include <react/renderer/components/view/ViewProps.h>
 #include "CompositionHelpers.h"
@@ -14,13 +15,24 @@ namespace Microsoft::ReactNative {
 struct CompositionBaseComponentView;
 struct CompContext;
 
+enum class CompositionComponentViewFeatures : std::uint_fast8_t {
+  None = 0,
+  NativeBorder = 1 << 0, // Standard border handling
+
+  Default = NativeBorder
+};
+
+DEFINE_ENUM_FLAG_OPERATORS(CompositionComponentViewFeatures);
+
 struct CompositionBaseComponentView : public IComponentView,
                                       public std::enable_shared_from_this<CompositionBaseComponentView> {
   static constexpr size_t SpecialBorderLayerCount = 8;
 
   CompositionBaseComponentView(
       const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-      facebook::react::Tag tag);
+      facebook::react::Tag tag,
+      winrt::Microsoft::ReactNative::ReactContext const &reactContext,
+      CompositionComponentViewFeatures flags);
 
   virtual winrt::Microsoft::ReactNative::Composition::IVisual Visual() const noexcept = 0;
   // Visual that should be parented to this ComponentView's parent
@@ -32,6 +44,11 @@ struct CompositionBaseComponentView : public IComponentView,
   const std::vector<IComponentView *> &children() const noexcept override;
   void parent(IComponentView *parent) noexcept override;
   IComponentView *parent() const noexcept override;
+  facebook::react::Props::Shared props() noexcept override;
+  virtual facebook::react::SharedViewProps viewProps() noexcept = 0;
+  void theme(const std::shared_ptr<Composition::Theme> &theme) noexcept override;
+  std::shared_ptr<Composition::Theme> &theme() const noexcept override;
+  void onThemeChanged() noexcept override;
   bool runOnChildren(bool forward, Mso::Functor<bool(IComponentView &)> &fn) noexcept override;
   void onFocusLost() noexcept override;
   void onFocusGained() noexcept override;
@@ -67,11 +84,15 @@ struct CompositionBaseComponentView : public IComponentView,
   facebook::react::Tag tag() const noexcept override;
 
   RECT getClientRect() const noexcept override;
+  facebook::react::Point getClientOffset() const noexcept override;
+  void updateProps(facebook::react::Props::Shared const &props, facebook::react::Props::Shared const &oldProps) noexcept
+      override;
+  void updateLayoutMetrics(
+      facebook::react::LayoutMetrics const &layoutMetrics,
+      facebook::react::LayoutMetrics const &oldLayoutMetrics) noexcept override;
+  void finalizeUpdates(RNComponentViewUpdateMask updateMask) noexcept override;
 
   void indexOffsetForBorder(uint32_t &index) const noexcept;
-  void updateBorderProps(
-      const facebook::react::ViewProps &oldViewProps,
-      const facebook::react::ViewProps &newViewProps) noexcept;
   void updateShadowProps(
       const facebook::react::ViewProps &oldViewProps,
       const facebook::react::ViewProps &newViewProps,
@@ -83,9 +104,6 @@ struct CompositionBaseComponentView : public IComponentView,
   void updateAccessibilityProps(
       const facebook::react::ViewProps &oldView,
       const facebook::react::ViewProps &newViewProps) noexcept;
-  void updateBorderLayoutMetrics(
-      facebook::react::LayoutMetrics const &layoutMetrics,
-      const facebook::react::ViewProps &viewProps) noexcept;
 
   virtual void OnRenderingDeviceLost() noexcept;
 
@@ -95,23 +113,16 @@ struct CompositionBaseComponentView : public IComponentView,
   void EnsureTransformMatrixFacade() noexcept;
 
   winrt::IInspectable EnsureUiaProvider() noexcept override;
+  std::optional<std::string> getAcccessiblityValue() noexcept override;
+  void setAcccessiblityValue(std::string &&value) noexcept override;
+  bool getAcccessiblityIsReadOnly() noexcept override;
+  ClipState getClipState() noexcept override;
 
   virtual std::string DefaultControlType() const noexcept;
   virtual std::string DefaultAccessibleName() const noexcept;
   virtual std::string DefaultHelpText() const noexcept;
 
  protected:
-  std::array<winrt::Microsoft::ReactNative::Composition::ISpriteVisual, SpecialBorderLayerCount>
-  FindSpecialBorderLayers() const noexcept;
-  bool TryUpdateSpecialBorderLayers(
-      std::array<winrt::Microsoft::ReactNative::Composition::ISpriteVisual, SpecialBorderLayerCount> &spBorderVisuals,
-      facebook::react::LayoutMetrics const &layoutMetrics,
-      const facebook::react::ViewProps &viewProps) noexcept;
-  void UpdateSpecialBorderLayers(
-      facebook::react::LayoutMetrics const &layoutMetrics,
-      const facebook::react::ViewProps &viewProps) noexcept;
-  void UpdateCenterPropertySet() noexcept;
-
   winrt::IInspectable m_uiaProvider{nullptr};
   winrt::Microsoft::ReactNative::Composition::ICompositionContext m_compContext;
   comp::CompositionPropertySet m_centerPropSet{nullptr};
@@ -119,13 +130,35 @@ struct CompositionBaseComponentView : public IComponentView,
   facebook::react::SharedViewEventEmitter m_eventEmitter;
   std::vector<IComponentView *> m_children;
   IComponentView *m_parent{nullptr};
+  RootComponentView *m_rootView{nullptr};
   facebook::react::LayoutMetrics m_layoutMetrics;
   bool m_needsBorderUpdate{false};
   bool m_hasTransformMatrixFacade{false};
   bool m_enableFocusVisual{false};
   uint8_t m_numBorderVisuals{0};
+  winrt::Microsoft::ReactNative::ReactContext m_context;
 
  private:
+  void updateBorderProps(
+      const facebook::react::ViewProps &oldViewProps,
+      const facebook::react::ViewProps &newViewProps) noexcept;
+  void updateBorderLayoutMetrics(
+      facebook::react::LayoutMetrics const &layoutMetrics,
+      const facebook::react::ViewProps &viewProps) noexcept;
+  void finalizeBorderUpdates(
+      facebook::react::LayoutMetrics const &layoutMetrics,
+      const facebook::react::ViewProps &viewProps) noexcept;
+  bool TryUpdateSpecialBorderLayers(
+      Composition::Theme &theme,
+      std::array<winrt::Microsoft::ReactNative::Composition::ISpriteVisual, SpecialBorderLayerCount> &spBorderVisuals,
+      facebook::react::LayoutMetrics const &layoutMetrics,
+      const facebook::react::ViewProps &viewProps) noexcept;
+  std::array<winrt::Microsoft::ReactNative::Composition::ISpriteVisual, SpecialBorderLayerCount>
+  FindSpecialBorderLayers() const noexcept;
+  void UpdateCenterPropertySet() noexcept;
+
+  CompositionComponentViewFeatures m_flags;
+  mutable std::shared_ptr<Composition::Theme> m_theme{nullptr};
   void showFocusVisual(bool show) noexcept;
   winrt::Microsoft::ReactNative::Composition::IFocusVisual m_focusVisual{nullptr};
   winrt::Microsoft::ReactNative::Composition::IVisual m_outerVisual{nullptr};
@@ -136,7 +169,8 @@ struct CompositionViewComponentView : public CompositionBaseComponentView {
 
   [[nodiscard]] static std::shared_ptr<CompositionViewComponentView> Create(
       const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-      facebook::react::Tag tag) noexcept;
+      facebook::react::Tag tag,
+      winrt::Microsoft::ReactNative::ReactContext const &reactContext) noexcept;
 
   void mountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
   void unmountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
@@ -147,7 +181,6 @@ struct CompositionViewComponentView : public CompositionBaseComponentView {
   void updateLayoutMetrics(
       facebook::react::LayoutMetrics const &layoutMetrics,
       facebook::react::LayoutMetrics const &oldLayoutMetrics) noexcept override;
-  void finalizeUpdates(RNComponentViewUpdateMask updateMask) noexcept override;
   void prepareForRecycle() noexcept override;
   bool focusable() const noexcept override;
   void onKeyDown(
@@ -158,7 +191,7 @@ struct CompositionViewComponentView : public CompositionBaseComponentView {
       const winrt::Microsoft::ReactNative::Composition::Input::KeyRoutedEventArgs &args) noexcept override;
   std::string DefaultControlType() const noexcept override;
 
-  facebook::react::Props::Shared props() noexcept override;
+  facebook::react::SharedViewProps viewProps() noexcept override;
 
   facebook::react::Tag hitTest(
       facebook::react::Point pt,
@@ -170,7 +203,8 @@ struct CompositionViewComponentView : public CompositionBaseComponentView {
  protected:
   CompositionViewComponentView(
       const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-      facebook::react::Tag tag);
+      facebook::react::Tag tag,
+      winrt::Microsoft::ReactNative::ReactContext const &reactContext);
 
  private:
   facebook::react::SharedViewProps m_props;