+++ vnext/Microsoft.ReactNative/Fabric/Composition/UnimplementedNativeViewComponentView.cpp
@@ -15,8 +15,13 @@ namespace Microsoft::ReactNative {
 
 UnimplementedNativeViewComponentView::UnimplementedNativeViewComponentView(
     const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-    facebook::react::Tag tag)
-    : Super(compContext, tag) {
+    facebook::react::Tag tag,
+    winrt::Microsoft::ReactNative::ReactContext const &reactContext)
+    : Super(
+          compContext,
+          tag,
+          reactContext,
+          (CompositionComponentViewFeatures::Default & ~CompositionComponentViewFeatures::NativeBorder)) {
   static auto const defaultProps = std::make_shared<facebook::react::UnimplementedNativeViewProps const>();
   m_props = defaultProps;
   m_visual = compContext.CreateSpriteVisual();
@@ -24,9 +29,10 @@ UnimplementedNativeViewComponentView::UnimplementedNativeViewComponentView(
 
 std::shared_ptr<UnimplementedNativeViewComponentView> UnimplementedNativeViewComponentView::Create(
     const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-    facebook::react::Tag tag) noexcept {
+    facebook::react::Tag tag,
+    winrt::Microsoft::ReactNative::ReactContext const &reactContext) noexcept {
   return std::shared_ptr<UnimplementedNativeViewComponentView>(
-      new UnimplementedNativeViewComponentView(compContext, tag));
+      new UnimplementedNativeViewComponentView(compContext, tag, reactContext));
 }
 
 void UnimplementedNativeViewComponentView::mountChildComponentView(
@@ -120,17 +126,17 @@ void UnimplementedNativeViewComponentView::updateLayoutMetrics(
       }
     }
   }
-  m_layoutMetrics = layoutMetrics;
+
+  Super::updateLayoutMetrics(layoutMetrics, oldLayoutMetrics);
 }
 
 void UnimplementedNativeViewComponentView::updateState(
     facebook::react::State::Shared const &state,
     facebook::react::State::Shared const &oldState) noexcept {}
 
-void UnimplementedNativeViewComponentView::finalizeUpdates(RNComponentViewUpdateMask updateMask) noexcept {}
-
 void UnimplementedNativeViewComponentView::prepareForRecycle() noexcept {}
-facebook::react::Props::Shared UnimplementedNativeViewComponentView::props() noexcept {
+
+facebook::react::SharedViewProps UnimplementedNativeViewComponentView::viewProps() noexcept {
   return m_props;
 }